Fix a problem to exceed int range.

+++ src/hashbucket.h
@@ -30,7 +30,7 @@ struct CombinerVal {
 };
 
 struct ReducerVal {
-    int   valbytes;
+    int64_t  valbytes;
     char *values_start;
     char *values_end;
 };
